Replace old SearchResult component
